Update type conversion from float to decimal in parsing
Replaced float.Parse with decimal.Parse to improve precision when parsing user input for transmission values. This ensures better accuracy and reduces potential rounding issues in the calculations.

+++ GT5 Car hack workshop 2.4/TransmissionEditor.cs
@@ -218,7 +218,7 @@ namespace GT5_Car_hack_workshop_2
         {
             try
             {
-                var working = (int)(float.Parse(textBoxValue) * 1000);
+                var working = (int)(decimal.Parse(textBoxValue) * 1000);
                 checked
                 {
                     MyProject.Forms.Form1.Gt5Save[moffOffset1] =
